Add sliced display mode.
Add a new display mode which renders transparent slices of the volume to
approximate Arnold 5 standard volume shader using single-scattering
equations.
The display mode uses a separate set of node attributes mimicking the
ones of the standard volume shader.

The slicing is done in an axis-aligned fashion by sampling the VDB grids
on a dense lattice. The volumes are cached up to a user-specified memory
limit to allow smooth animation of volumes.